requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mongoengine==0.27.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pymongo==4.5.0
dnspython==2.4.2
cloudinary==1.36.0
